+++ clients/odbc/driver/SQLGetInfo.c
@@ -1576,7 +1576,7 @@ SQLGetInfo(SQLHDBC ConnectionHandle,
 #ifdef ODBCDEBUG
 	ODBCLOG("SQLGetInfo " PTRFMT " %s " PTRFMT " %d " PTRFMT "\n",
 		PTRFMTCAST ConnectionHandle, translateInfoType(InfoType),
-		PTRFMTCAST InfoValuePtr, BufferLength, 
+		PTRFMTCAST InfoValuePtr, (int) BufferLength,
 		PTRFMTCAST StringLengthPtr);
 #endif
 
@@ -1621,7 +1621,7 @@ SQLGetInfoW(SQLHDBC ConnectionHandle,
 #ifdef ODBCDEBUG
 	ODBCLOG("SQLGetInfoW " PTRFMT " %s " PTRFMT " %d " PTRFMT "\n",
 		PTRFMTCAST ConnectionHandle, translateInfoType(InfoType),
-		PTRFMTCAST InfoValuePtr, BufferLength, 
+		PTRFMTCAST InfoValuePtr, (int) BufferLength,
 		PTRFMTCAST StringLengthPtr);
 #endif
